GENERAL- CUSTOMER SUPPLIER ADDED

+++ Dfinance/Dfinance.Core/Views/Common/NextCodeView.cs
@@ -8,7 +8,8 @@ namespace Dfinance.Core.Views.Common
 {
     public class NextCodeView
     {
-        public long Code {  get; set; }
+        public int Code {  get; set; }
+
     }
     public class AccountCodeView 
     {
